Cleaning Up
Implementing AJAX to avoid the need for page refreshes. Posting works
without refresh. Buttons and all elements are drawn dynamically without
refresh.

UI cleaned up. Added a scrollbar to the unpublished tweets section to
add consistency to the UI (a paging feature would also work, but the
scrollbar is only 2 lines of CSS code).

Need to implement AJAX functions for up-vote, down-vote, edit, and
publish (this update disables these functions — buttons will only send
an alert to the user of the function type and post ID related to the
button clicked). This should be relatively simple considering a
majority of the back-end code has already been completed.